TP3 - long long for perfs

+++ TP3/perf.h
@@ -13,6 +13,6 @@ void perf_printh(const perf_t * p);
 
 void perf_printmicro(const perf_t * p);
 
-double perf_mflops(const perf_t * p, const long nb_op);
+double perf_mflops(const perf_t * p, const long long nb_op);
 
 #endif //TDP_PERF_H
